feat: 전체적인 웹 스타일 적용 및 사용자 경험 개선을 위한 유틸 기능 추가 - 로딩화면을 보여줄 시 특정 시간안에 데이터가 로드되면 로딩스피너를 보여주지 않도록 적용 - App 컴포넌트 레벨에서 페이지가 네비게이션 될 시 스크롤바 최상단에 위치시키도록 함 - 방문했던 검색결과과 게시물 표지를 세션스토리지에 일정시간 저장함. 페이지 네비게이터 사용시 일정시간동안 페이지를 기억하도록 세션 스토리지에 저장함 - 게시물 생성시 서버에서 요구하는 글자수를 초과시 사용자에게 알려주도록 - 내가 만든 게시물 상세 페이지에서 게시물 링크를 공유할 수 있도록 링크 복사 기능 적용

+++ src/util/storage.ts
@@ -3,13 +3,13 @@ interface IStorageData<T> {
   timeStamp: number;
 }
 
-type TSaveItemInSession = 'success' | 'security' | 'unknown';
-
 // 검색 결과, 판도라 표지의 데이터 신선도. 이 기간동안은 세션스토리지 데이터를 이용함
 const EXPIRATION_TIME = 1 * 60 * 1000;
 
-export const saveInSession = <T>(key: string, data: T): TSaveItemInSession => {
+export const saveInSession = <T>(key: string, data: T): void => {
   try {
+    console.log(key);
+    console.log(data);
     const now = Date.now();
     const storageData: IStorageData<T> = {
       value: data,
@@ -17,18 +17,12 @@ export const saveInSession = <T>(key: string, data: T): TSaveItemInSession => {
     };
 
     sessionStorage.setItem(key, JSON.stringify(storageData));
-    return 'success';
   } catch (error) {
-    if (error instanceof DOMException) {
-      if (error.name === 'QuotaExceededError') {
-        sessionStorage.clear();
-      }
-      if (error.name === "SecurityError") {
-        return 'security';
-      }
-      return 'unknown';
+    if (error instanceof DOMException && error.name === 'QuotaExceededError') {
+      console.log('seesion error');
+      sessionStorage.clear();
     }
-    return 'unknown';
+    // 에러 다이렉트 만들기
   }
 }
